oop_patterns/requirements.txt
numpy==1.16.1
PyYAML==3.12
pygame==2.0.1